pyproject.toml
[project]
name = "order-guard"
version = "0.1.0"
description = "企业数据智能监控中台"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "litellm>=1.55.0",
    "apscheduler>=4.0.0a5",
    "httpx>=0.28.0",
    "typer>=0.15.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "aiosqlite>=0.20.0",
    "greenlet>=3.3.2",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "chardet>=5.2.0",
    "mcp>=1.0.0",
    "sqlglot>=26.0.0",
    "lark-oapi>=1.5.3",
    "tabulate>=0.10.0",
    "croniter>=2.0.0",
    "kaggle>=2.0.0",
    "mysql-connector-python>=9.6.0",
    "langwatch-scenario>=0.7.21",
]

[project.scripts]
order-guard = "order_guard.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/order_guard"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "e2e: End-to-end tests requiring LLM API key",
]
